requirements.txt
amqp==5.1.1
APScheduler==3.6.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
cachetools==4.2.2
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
Deprecated==1.2.13
dj-database-url==1.0.0
Django==3.2.9
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-debug-toolbar==3.6.0
django-timezone-field==5.0
executing==1.2.0
google-api-core==2.11.0
google-api-python-client==2.74.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
gspread==5.7.2
gunicorn==20.1.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
ipython==8.5.0
jedi==0.18.2
kombu==5.2.4
matplotlib-inline==0.1.6
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
protobuf==4.21.12
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-telegram-bot==13.11
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tornado==6.2
traitlets==5.8.1
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.18.3
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.2.0
wrapt==1.14.1
PySocks==1.7.1
